addding delete review and add review funcitonality

+++ client/src/ReviewCard.js
@@ -1,28 +1,33 @@
-import React, { useState } from "react";
+import React, { useState, useContext } from "react";
 import EditReviewForm from "./EditReviewForm";
+import { UserContext } from "./context/user";
 
 function ReviewCard({ review, deleteReview }) {
+  const { user } = useContext(UserContext);
   const [showEditForm, setShowEditForm] = useState(false);
   const [rev, setRev] = useState(review);
 
   function handleDeleteReviewClick() {
-    fetch(`/reviews/${rev.id}`, {
-      method: "DELETE",
-    }).then((res) => {
-      if (res.ok) {
-        deleteReview(rev.id);
-      }
-    });
+    if (user.id === rev.user_id) {
+      fetch(`/reviews/${rev.id}`, {
+        method: "DELETE",
+      }).then((res) => {
+        if (res.ok) {
+          deleteReview(rev.id);
+        }
+      });
+    }
   }
 
   function toggleEditReviewForm() {
-    setShowEditForm(!showEditForm);
+    if (user.id === rev.user_id) {
+      setShowEditForm(!showEditForm);
+    }
   }
 
   return (
     <div key={rev.id}>
       <h3>{rev.title}</h3>
-      <button>Edit</button>
       {showEditForm ? (
         <EditReviewForm
           review={rev}
@@ -32,6 +37,7 @@ function ReviewCard({ review, deleteReview }) {
       ) : (
         <button onClick={toggleEditReviewForm}>Edit</button>
       )}
+
       <button onClick={handleDeleteReviewClick}>Delete</button>
       <br />
       {rev.body}
